pyproject.toml
[project]
name = "autoopsai"
version = "0.2.0"
description = "AI-powered virtual DevOps engineer with human-in-the-loop approval"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.42.0",
    "websockets>=15.0",
    # LLM
    "ollama>=0.4.0",
    # Database
    "motor>=3.6.0",
    "beanie>=1.27.0,<2.0",
    "pymongo>=4.9.0,<4.10",
    # HTTP / resilience
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    # Models / config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Infrastructure
    "docker>=7.1.0",
    "pyyaml>=6.0.3",
    # Observability
    "prometheus-client>=0.22.0",
    # Utilities
    "rich>=13.0.0",
    "watchfiles>=1.0.0",
    "python-dotenv>=1.2.2",
    # Teams (POC only — production Teams deferred)
    "botbuilder-core>=4.17.1",
    "botbuilder-integration-aiohttp>=4.17.1",
    "aiohttp>=3.13.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.11.0",
]
